aider: Refactored `pyyt.py` by adding comments (docstrings) and moving methods to other files (like `utils.py`) where necessary.
# Context:
USER: refactor the file `pyyt.py` by adding comments (docstrings) and moving methods to other files (like for example utils.py) where necesasry
