+++ pyyt.py
@@ -4,45 +4,13 @@ import pathlib
 import sys
 
 import yt_dlp
-
-ydl_opts = {
-    "format": "bestaudio/best",
-    "postprocessors": [
-        {
-            "key": "FFmpegExtractAudio",
-            "preferredcodec": "mp3",
-            "preferredquality": "192",
-        },
-    ],
-    "outtmpl": f"{pathlib.Path().absolute()}/downloads/%(title)s.%(ext)s",
-    "extract_flat": "in_playlist",
-    "ignoreerrors": True,
-    "quiet": False,
-    "sleep_interval": 2,
-}
-ydl = yt_dlp.YoutubeDL(ydl_opts)
-
-
-def get_video_entries(playlist_url):
-    results = ydl.extract_info(
-        playlist_url,
-        download=False,  # We just want to extract the info
-    )
-
-    if "entries" in results:
-        entries = results["entries"]
-        return entries
-
-    print("Something went wrong, no entries in playlist...")
-    sys.exit(2)
-
-
-def download_and_metadata(video_url):
-    # TODO add metadata handling
-    ydl.download([video_url])
+from utils import get_video_entries, download_and_metadata
 
 
 def main():
+    """
+    Main function to run the pyyt script.
+    """
     print("Welcome to pyyt - Select your option:\n")
     print("1 - Download an entire playlist as audio\n")
     print("2 - Download a single video as audio\n")
